bchanged details
