+++ app.js
@@ -19,6 +19,7 @@ app.use((req, res) => {
   res.status(404).json({ message: 'Not found' })
 })
 app.use((err, req, res, next) => {
-  res.status(500).json({ message: err.message })
+  const {status,message} = err
+  res.status(status).json({message})
 })
 module.exports = app
