+++ routes/api/auth.js
@@ -1,14 +1,15 @@
 const express = require("express");
 const router = express.Router();
-const { validateBody, authenticate, uploadAvatar } = require("../../services/middleware.js");
+const { validateBody, authenticate} = require("../../services/middleware.js");
 const { schemas } = require("../../services/validationUser.js");
-const { regist, login, logOut } = require("../../services/auth.js");
+const { regist, login, logOut, currentUser } = require("../../services/auth.js");
 const { upload } = require("../../services/upload.js");
+const { uploadAvatar } = require("../../services/uploadAvatar.js");
 
-router.post("/auth/register", validateBody(schemas.registSchema), regist);
-router.post("/auth/login", validateBody(schemas.loginSchema), login);
-router.post("/auth/logout",authenticate,logOut);
-router.get("/current",authenticate,login);
+router.post("/register", validateBody(schemas.registSchema), regist);
+router.post("/login", validateBody(schemas.loginSchema), login);
+router.post("/logout",authenticate,logOut);
+router.get("/current",authenticate,currentUser);
 router.patch("/avatars",authenticate, upload.single("avatar"), uploadAvatar);
